move stuff to lib and fix comment internal assert
