+++ lib/errors.js
