+++ lib/snapshot.js
@@ -38,7 +38,7 @@ exports.getSnapshot = function (filename, key, actual) {
     if (isTypedArray(v)) {
       if (!brittle) {
         brittle = true
-        s = 'const { createTypedArray } = require(\'brittle/snapshot\')\n\n' + s
+        s = 'const { createTypedArray } = require(\'brittle\')\n\n' + s
       }
       const b = b4a.from(v.buffer, v.byteOffset, v.byteLength)
       s += 'createTypedArray(' + v.constructor.name + ', ' + toString(b4a.toString(b, 'base64')) + ')\n\n'
